refactor: 合并模板变量选择菜单到代码编辑器设置卡
